+++ packages/design-system/src/components/index.ts
@@ -2,7 +2,6 @@ export * from './Accordion'
 export * from './AiBadge'
 export * from './Avatar'
 export * from './Button'
-export * from './ButtonLink'
 export * from './Dialog'
 export * from './Drawer'
 export * from './GenericPlaceholder'
