+++ src/components/subscriptions/SubscriptionAlertsList.tsx
@@ -1,9 +1,9 @@
 import { gql } from '@apollo/client'
-import { Button, ButtonLink, Icon, Typography } from 'lago-design-system'
+import { Button, Icon, Typography } from 'lago-design-system'
 import { useCallback, useRef } from 'react'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
 
-import { Table } from '~/components/designSystem/Table'
+import { ButtonLink, Table } from '~/components/designSystem'
 import {
   DeleteAlertDialog,
   DeleteAlertDialogRef,
