+++ src/layouts/SettingsNavLayout.tsx
@@ -1,5 +1,4 @@
 import { ClickAwayListener } from '@mui/material'
-import { ButtonLink } from 'lago-design-system'
 import { useEffect, useRef, useState } from 'react'
 import {
   generatePath,
@@ -12,6 +11,7 @@ import {
 
 import {
   Button,
+  ButtonLink,
   Skeleton,
   Typography,
   VerticalMenu,
