+++ src/pages/subscriptions/SubscriptionDetails.tsx
@@ -1,5 +1,5 @@
 import { gql } from '@apollo/client'
-import { Avatar, ButtonLink, Icon } from 'lago-design-system'
+import { Avatar, Icon } from 'lago-design-system'
 import { useMemo, useRef } from 'react'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
 
@@ -7,7 +7,14 @@ import {
   TerminateCustomerSubscriptionDialog,
   TerminateCustomerSubscriptionDialogRef,
 } from '~/components/customers/subscriptions/TerminateCustomerSubscriptionDialog'
-import { Button, NavigationTab, Popper, Skeleton, Typography } from '~/components/designSystem'
+import {
+  Button,
+  ButtonLink,
+  NavigationTab,
+  Popper,
+  Skeleton,
+  Typography,
+} from '~/components/designSystem'
 import { DetailsPage } from '~/components/layouts/DetailsPage'
 import { SubscriptionActivityLogs } from '~/components/subscriptions/SubscriptionActivityLogs'
 import { SubscriptionAlertsList } from '~/components/subscriptions/SubscriptionAlertsList'
